add product to cart

+++ database/migrations/2018_04_17_190701_create_transactions_table.php
@@ -17,7 +17,7 @@ class CreateTransactionsTable extends Migration
             $table->increments('id');
             $table->integer('user_id');
             $table->integer('customer_id');
-            $table->float('total');
+            $table->double('total');
             $table->timestamps();
         });
     }
